+++ src/wf.rs
@@ -17,13 +17,20 @@ use crate::{DecodeError, FieldInfo};
 
 /// Decodes the ISS value for a trapped WF* instruction.
 pub fn decode_iss_wf(iss: u64) -> Result<Vec<FieldInfo>, DecodeError> {
-    let cv = FieldInfo::get_bit(iss, "CV", 24).describe_bit(describe_cv);
-    let cond = FieldInfo::get(iss, "COND", 20, 24);
-    let res0a = FieldInfo::get(iss, "RES0", 10, 20).check_res0()?;
-    let rn = FieldInfo::get(iss, "RN", 5, 10);
-    let res0b = FieldInfo::get(iss, "RES0", 3, 5).check_res0()?;
-    let rv = FieldInfo::get_bit(iss, "RV", 2).describe_bit(describe_rv);
-    let ti = FieldInfo::get(iss, "TI", 0, 2).describe(describe_ti)?;
+    let cv =
+        FieldInfo::get_bit(iss, "CV", Some("Condition code valid"), 24).describe_bit(describe_cv);
+    let cond = FieldInfo::get(
+        iss,
+        "COND",
+        Some("Condition code of the trapped instruction"),
+        20,
+        24,
+    );
+    let res0a = FieldInfo::get(iss, "RES0", Some("Reserved"), 10, 20).check_res0()?;
+    let rn = FieldInfo::get(iss, "RN", Some("Register Number"), 5, 10);
+    let res0b = FieldInfo::get(iss, "RES0", Some("Reserved"), 3, 5).check_res0()?;
+    let rv = FieldInfo::get_bit(iss, "RV", Some("Register Valid"), 2).describe_bit(describe_rv);
+    let ti = FieldInfo::get(iss, "TI", Some("Trapped Instruction"), 0, 2).describe(describe_ti)?;
 
     Ok(vec![cv, cond, res0a, rn, res0b, rv, ti])
 }